Add files via upload
It is hotel management kind of theme project but totally different from normal one as It is like place where we can listen song/dialogue given by hotel as per individual choice and also provide one game for playing it till your order arrives. So it is helpful for all to remove the boring time period between ordering food to arriving food. So that this kind of new idea can change your point of view during any kind of real-time problem